* Add new column profile's users list * Add grouped delete task to profiles' management table.
git-svn-id: file:///home/code/svn/ovd/trunk@7776 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/profiles.php
@@ -4,6 +4,7 @@
  * http://www.ulteo.com
  * Author Laurent CLOUET <laurent@ulteo.com> 2010-2011
  * Author Jeremy DESVAGES <jeremy@ulteo.com> 2009
+ * Author Omar AKHAM <oakham@ulteo.com> 2011
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -49,24 +50,78 @@ function show_default() {
 	echo '<div id="profiles_div">';
 	echo '<h1>'._('Profiles').'</h1>';
 
-	echo '<div id="profiles_list_div">';
-	echo '<table border="0" cellspacing="1" cellpadding="3">';
+	$nb_profiles = count($profiles);
 
-	foreach ($profiles as $profile) {
-		echo '<tr>';
-		echo '<td><a href="profiles.php?action=manage&amp;id='.$profile->id.'">'.$profile->id.'</a></td>';
-		if ($can_manage_profiles) {
-			echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this profile?').'\');">';
+	if($nb_profiles > 0) {
+		echo '<div id="profiles_list_div">';
+		echo '<table id="profils_list_table" border="0" class="main_sub sortable" cellspacing="1" cellpadding="3">';	
+		echo '<thead>';
+		echo '<tr class="title">';
+		if ($nb_profiles > 1 and $can_manage_profiles)
+			echo '<th class="unsortable"></th>';
+		echo '<th>'._('Profile').'</th><th>'._('User').'</th>';
+		echo '</tr>';
+		echo '</thead>';
+		echo '<tbody>';
+		$count = 0;
+		
+		foreach ($profiles as $profile) {
+			$content = 'content'.(($count++%2==0)?1:2);
+			echo '<tr class="'.$content.'">';
+			
+			if($nb_profiles > 1 and $can_manage_profiles)
+				echo '<td><input class="input_checkbox" type="checkbox" name="ids[]" value="'.$profile->id.'" /></td>';
+			
+			echo '<td><a href="profiles.php?action=manage&amp;id='.$profile->id.'">'.$profile->id.'</a></td>';
+			$users = '';
+			$profile_users = $profile->getUsers();
+			$count1 = count($profile_users);
+			
+			foreach($profile_users as $user)
+				$users[] = '<a href="users.php?action=manage&amp;id='.$user->getAttribute('login').'">'.$user->getAttribute('displayname').'</a>';
+			
+			$str_users = implode(', ', $users);
+			echo '<td>'.$str_users.'</td>';
+			
+			if ($can_manage_profiles) {
+				echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this profile?').'\');">';
+				echo '<input type="hidden" name="name" value="Profile" />';
+				echo '<input type="hidden" name="action" value="del" />';
+				echo '<input type="hidden" name="ids[]" value="'.$profile->id.'" />';
+				echo '<input type="submit" value="'._('Delete this profile').'" />';
+				echo '</form></td>';
+			}
+			
+			echo '</tr>';
+		}
+		
+		echo '</tbody>';
+		
+		if ($nb_profiles > 1 and $can_manage_profiles) {
+			$content = 'content'.(($count++%2==0)?1:2);
+			echo '<tfoot>';
+			echo '<tr class="'.$content.'">';
+			echo '<td colspan="'.(($nb_profiles > 0)?3:2).'">';
+			echo '<a href="javascript:;" onclick="markAllRows(\'profils_list_table\'); return false">'._('Mark all').'</a>';
+			echo ' / <a href="javascript:;" onclick="unMarkAllRows(\'profils_list_table\'); return false">'._('Unmark all').'</a>';
+			echo '</td>';
+			echo '<td>';
+			echo '<form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete selected profiles?').'\') && updateMassActionsForm(this, \'profils_list_table\');">';
 			echo '<input type="hidden" name="name" value="Profile" />';
 			echo '<input type="hidden" name="action" value="del" />';
-			echo '<input type="hidden" name="ids[]" value="'.$profile->id.'" />';
-			echo '<input type="submit" value="'._('Delete this profile').'" />';
-			echo '</form></td>';
+			echo '<input type="submit" value="'._('Delete profiles').'"/>';
+			echo '</form>';
+			echo '</td>';
+			echo '</tr>';
+			echo '</tfoot>';
 		}
-		echo '</tr>';
+		
+		echo '</table>';
+	}
+	else {
+		echo _('No profile');
 	}
 
-	echo '</table>';
 	echo '</div>';
 
 	echo '</div>';
